chore: reducer를 moneyReducer와 recordReducer로 분리
