+++ src/components/recordBoard/RecordBoard.jsx
@@ -3,11 +3,11 @@ import { useVendingMachineContext } from "../../context/VendingMachineContext";
 import { RecordContainer } from "./RecordBoard.style";
 
 function RecordBoard() {
-    const { money } = useVendingMachineContext();
+    const { record } = useVendingMachineContext();
 
     return (
         <RecordContainer>
-            {money.record.map((r) => (
+            {record.map((r) => (
                 <p key={uuid()}>{r}</p>
             ))}
         </RecordContainer>
